requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
yowsup2==2.5.2
requests==2.9.1
bs4==0.0.1
forex-python==0.3.2
lxml==3.7.3
demjson==2.2.4
pycurl==7.43.0
#pytube==6.2.2
pytube==7.0.18
pywhatsapp==0.0.7
pymysql==0.7.11
simplejson==3.11.1
beckett==0.8.0
blinker==1.4.0
#pytube
Pillow==4.3.0
translate==2.0.0
emoji==0.4.5
#av==0.3.3
pyfiglet==0.7.5
geopy==1.11.0
python-forecastio==1.3.5
pytz==2017.3
tox==2.9.1
preview-generator==0.2.1
audioread==2.1.4
wkhtmltopdf==0.2
gTTS==1.2.2
you-get==0.4.985
goslate==1.5.1
#ChatterBot==0.8.1
amzsear==1.0.5
pytelegrambotapi==3.6.3